requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
WTForms==3.0.1
itsdangerous==2.1.2
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0